* Makefile.in (check-unix): Built and execture dyntest
	* test.c: Include string,h, stdlib.h.

	* dyn_create.c, dyn_delete.c, dyn_insert.c, dyn_put.c,
	dyn_realloc.c: Cast arguments to malloc(), realloc(), memmove() to
	size_t.

	* dynP.h: Provide full prototypes for _DynRealloc() and _DynResize().

	* dyn.h: Add prototype for DynAppend.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12866 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/dyn/dyn_delete.c
@@ -40,7 +40,7 @@ int DynDelete(obj, idx)
 	  if (obj->paranoid) {
 	       if (obj->debug)
 		    fprintf(stderr, "dyn: delete: last element, zeroing.\n");
-	       memset(obj->array + idx*obj->el_size, 0, obj->el_size);
+	       memset(obj->array + idx*obj->el_size, 0, (size_t) obj->el_size);
 	  }
 	  else {
 	       if (obj->debug)
@@ -57,7 +57,7 @@ int DynDelete(obj, idx)
 #ifdef HAVE_MEMMOVE
 	  memmove(obj->array + idx*obj->el_size,
 		  obj->array + (idx+1)*obj->el_size,
-		  obj->el_size*(obj->num_el - idx));
+		  (size_t) obj->el_size*(obj->num_el - idx));
 #else
 	  bcopy(obj->array + (idx+1)*obj->el_size,
 		  obj->array + idx*obj->el_size,
@@ -70,7 +70,7 @@ int DynDelete(obj, idx)
 			    obj->el_size, obj->array,
 			    obj->el_size*(obj->num_el - 1));
 	       memset(obj->array + obj->el_size*(obj->num_el - 1), 0,
-		     obj->el_size);
+		     (size_t) obj->el_size);
 	  }
      }
